+++ crates/core/task/Cargo.toml
@@ -14,3 +14,6 @@ rustc-hash = { workspace = true }
 serde = { workspace = true }
 strum = { version = "0.24.1", features = ["derive"] }
 thiserror = { workspace = true }
+
+[dev-dependencies]
+moon_test_utils = { path = "../test-utils" }
