+++ crates/core/task/src/lib.rs
@@ -6,10 +6,9 @@ pub mod test;
 mod token;
 mod types;
 
-pub use moon_config::{PlatformType, TargetID, TaskConfig, TaskID, TaskOptionsConfig};
-
 pub use errors::*;
 pub use file_group::FileGroup;
+pub use moon_config::{PlatformType, TargetID, TaskConfig, TaskID, TaskOptionsConfig};
 pub use target::{Target, TargetProjectScope};
 pub use task::{Task, TaskOptions};
 pub use token::{ResolverData, ResolverType, TokenResolver, TokenType};
