+++ crates/core/task/tests/task_test.rs
@@ -1,7 +1,7 @@
 use moon_config::{TaskCommandArgs, TaskConfig, TaskOptionEnvFile, TaskOptionsConfig};
 use moon_task::test::create_expanded_task;
 use moon_task::{Target, Task, TaskOptions};
-use moon_utils::test::{create_sandbox, get_fixtures_dir};
+use moon_test_utils::{create_sandbox, get_fixtures_dir};
 use moon_utils::{glob, string_vec};
 use rustc_hash::{FxHashMap, FxHashSet};
 use std::env;
@@ -24,9 +24,8 @@ fn errors_for_output_glob() {
 }
 
 mod from_config {
-    use moon_config::{TaskMergeStrategy, TaskOutputStyle};
-
     use super::*;
+    use moon_config::{TaskMergeStrategy, TaskOutputStyle};
 
     #[test]
     fn sets_defaults() {
@@ -510,13 +509,13 @@ mod expand_env {
     #[test]
     #[should_panic(expected = "Error parsing line: 'FOO', error at line index: 3")]
     fn errors_on_invalid_file() {
-        let fixture = create_sandbox("cases");
-        let project_root = fixture.path().join("base");
+        let sandbox = create_sandbox("cases");
+        let project_root = sandbox.path().join("base");
 
         fs::write(project_root.join(".env"), "FOO").unwrap();
 
         create_expanded_task(
-            fixture.path(),
+            sandbox.path(),
             &project_root,
             Some(TaskConfig {
                 options: TaskOptionsConfig {
@@ -538,11 +537,11 @@ mod expand_env {
         if moon_utils::is_ci() {
             panic!("InvalidEnvFile");
         } else {
-            let fixture = create_sandbox("cases");
-            let project_root = fixture.path().join("base");
+            let sandbox = create_sandbox("cases");
+            let project_root = sandbox.path().join("base");
 
             create_expanded_task(
-                fixture.path(),
+                sandbox.path(),
                 &project_root,
                 Some(TaskConfig {
                     options: TaskOptionsConfig {
@@ -558,13 +557,13 @@ mod expand_env {
 
     #[test]
     fn loads_using_bool() {
-        let fixture = create_sandbox("cases");
-        let project_root = fixture.path().join("base");
+        let sandbox = create_sandbox("cases");
+        let project_root = sandbox.path().join("base");
 
         fs::write(project_root.join(".env"), "FOO=foo\nBAR=123").unwrap();
 
         let task = create_expanded_task(
-            fixture.path(),
+            sandbox.path(),
             &project_root,
             Some(TaskConfig {
                 options: TaskOptionsConfig {
@@ -587,13 +586,13 @@ mod expand_env {
 
     #[test]
     fn loads_using_custom_path() {
-        let fixture = create_sandbox("cases");
-        let project_root = fixture.path().join("base");
+        let sandbox = create_sandbox("cases");
+        let project_root = sandbox.path().join("base");
 
         fs::write(project_root.join(".env.production"), "FOO=foo\nBAR=123").unwrap();
 
         let task = create_expanded_task(
-            fixture.path(),
+            sandbox.path(),
             &project_root,
             Some(TaskConfig {
                 options: TaskOptionsConfig {
@@ -616,13 +615,13 @@ mod expand_env {
 
     #[test]
     fn doesnt_override_other_env() {
-        let fixture = create_sandbox("cases");
-        let project_root = fixture.path().join("base");
+        let sandbox = create_sandbox("cases");
+        let project_root = sandbox.path().join("base");
 
         fs::write(project_root.join(".env"), "FOO=foo\nBAR=123").unwrap();
 
         let task = create_expanded_task(
-            fixture.path(),
+            sandbox.path(),
             &project_root,
             Some(TaskConfig {
                 env: Some(FxHashMap::from_iter([(
