+++ crates/core/task/tests/token_test.rs
@@ -1,7 +1,7 @@
 use moon_config::{ProjectConfig, ProjectLanguage, ProjectType, TaskConfig};
 use moon_task::test::{create_expanded_task, create_file_groups, create_initial_task};
 use moon_task::{ResolverData, TokenResolver};
-use moon_utils::test::get_fixtures_dir;
+use moon_test_utils::get_fixtures_dir;
 use moon_utils::{glob, string_vec};
 use std::path::PathBuf;
